refactor: IntegrationTest 패키지 변경
